refactored basxconnect.core.views
