+++ basxconnect/core/views/__init__.py
@@ -1,0 +1,5 @@
+from .menu_views import *  # noqa
+from .person.person_browse_views import *  # noqa
+from .person.person_details_views import *  # noqa
+from .person.person_search_views import *  # noqa
+from .settings_views import *  # noqa
