+++ basxconnect/core/views/person/person_browse_views.py
@@ -1,105 +1,20 @@
 import htmlgenerator as hg
 from bread import layout as layout
-from bread import menu
-from bread.forms.forms import generate_form
 from bread.layout.components.datatable import DataTableColumn
 from bread.menu import Link
-from bread.utils.urls import aslayout, reverse, reverse_model
-from bread.views import BrowseView, EditView, ReadView, layoutasreadonly
+from bread.utils.urls import reverse, reverse_model
+from bread.views import BrowseView
 from django import forms
 from django.db.models import Q
-from django.http import HttpResponse
-from django.shortcuts import get_object_or_404
 from django.utils.html import mark_safe
 from django.utils.translation import gettext_lazy as _
 from django.utils.translation import pgettext_lazy
-from django.views.decorators.csrf import csrf_exempt
-from haystack.query import SearchQuerySet
-from haystack.utils.highlighting import Highlighter
 
-import basxconnect.core.layouts.settings_layout as settings_layout
-from basxconnect.core.layouts.editlegalperson import editlegalperson_form
-from basxconnect.core.layouts.editnaturalperson import editnaturalperson_form
-from basxconnect.core.layouts.editperson import editperson_head, editperson_toolbar
-from basxconnect.core.layouts.editpersonassociation import editpersonassociation_form
-
-from . import models, settings
+from basxconnect.core import models, settings
 
 R = layout.grid.Row
 C = layout.grid.Col
 F = layout.form.FormField
-
-# ADD MODEL VIEWS AND REGISTER URLS -------------------------------------------
-
-
-def personform_shortcut(request, formlayout, isreadview):
-    return hg.BaseElement(
-        layout.grid.Grid(
-            editperson_toolbar(request),
-            editperson_head(request, isreadview=isreadview),
-            layout.form.Form(hg.C("form"), formlayout),
-        )
-    )
-
-
-class NaturalPersonEditView(EditView):
-    def get_layout(self):
-        return personform_shortcut(
-            self.request,
-            editnaturalperson_form(self.request),
-            isreadview=False,
-        )
-
-
-class NaturalPersonReadView(ReadView):
-    def get_layout(self):
-        return layoutasreadonly(
-            personform_shortcut(
-                self.request,
-                editnaturalperson_form(self.request),
-                isreadview=True,
-            )
-        )
-
-
-class LegalPersonEditView(EditView):
-    def get_layout(self):
-        return personform_shortcut(
-            self.request,
-            editlegalperson_form(self.request),
-            isreadview=False,
-        )
-
-
-class LegalPersonReadView(ReadView):
-    def get_layout(self):
-        return layoutasreadonly(
-            personform_shortcut(
-                self.request,
-                editlegalperson_form(self.request),
-                isreadview=True,
-            )
-        )
-
-
-class PersonAssociationEditView(EditView):
-    def get_layout(self):
-        return personform_shortcut(
-            self.request,
-            editpersonassociation_form(self.request),
-            isreadview=False,
-        )
-
-
-class PersonAssociationReadView(ReadView):
-    def get_layout(self):
-        return layoutasreadonly(
-            personform_shortcut(
-                self.request,
-                editpersonassociation_form(self.request),
-                isreadview=True,
-            )
-        )
 
 
 class PersonBrowseView(BrowseView):
@@ -440,170 +355,3 @@ class PersonBrowseView(BrowseView):
                 columns.append("preferred_language")
 
         return super().export(*args, columns=columns, **kwargs)
-
-
-# ADD SETTING VIEWS AND REGISTER URLS -------------------------------------------
-
-
-@aslayout
-def generalsettings(request):
-    layoutobj = settings_layout.generalsettings(request)
-    form = None
-    if models.Person.objects.filter(id=settings.OWNER_PERSON_ID).exists():
-        form = generate_form(
-            request,
-            models.LegalPerson,
-            layoutobj,
-            models.Person.objects.filter(id=settings.OWNER_PERSON_ID).first(),
-        )
-
-        if request.method == "POST":
-            if form.is_valid():
-                form.save()
-
-        content = layout.form.Form(form, layoutobj, style="max-width: 480px")
-    else:
-        content = (
-            _(
-                "The django setting BASXCONNECT.OWNER_PERSON_ID needs to be set to an existing person in order to be able to edit this screen"
-            ),
-        )
-    return hg.WithContext(
-        layout.grid.Grid(
-            R(C(hg.H3(_("Settings")))),
-            R(C(hg.H4(_("General")))),
-            R(C(hg.H5(_("Information about our organization")))),
-            R(C(content)),
-        ),
-        form=form,
-    )
-
-
-@csrf_exempt
-def togglepersonstatus(request, pk: int):
-    if request.method == "POST":
-        person = get_object_or_404(models.Person, pk=pk)
-        person.active = not person.active
-        person.save()
-    return HttpResponse(
-        _("%s is %s") % (person, _("Active") if person.active else _("Inactive"))
-    )
-
-
-@aslayout
-def personsettings(request):
-    return settings_layout.personsettings(request)
-
-
-@aslayout
-def relationshipssettings(request):
-    return settings_layout.relationshipssettings(request)
-
-
-# MENU ENTRIES ---------------------------------------------------------------------
-
-persongroup = menu.Group(_("Persons"), icon="group")
-settingsgroup = menu.Group(_("Settings"), icon="settings", order=100)
-
-menu.registeritem(
-    menu.Item(
-        menu.Link(reverse_model(models.Person, "browse"), _("Persons")), persongroup
-    )
-)
-
-menu.registeritem(
-    menu.Item(
-        menu.Link(reverse("basxconnect.core.views.generalsettings"), _("General")),
-        settingsgroup,
-    )
-)
-menu.registeritem(
-    menu.Item(
-        menu.Link(reverse("basxconnect.core.views.personsettings"), _("Persons")),
-        settingsgroup,
-    )
-)
-menu.registeritem(
-    menu.Item(
-        menu.Link(
-            reverse("basxconnect.core.views.relationshipssettings"),
-            _("Relationships"),
-        ),
-        settingsgroup,
-    )
-)
-
-
-class CustomHighlighter(Highlighter):
-    def find_window(self, highlight_locations):
-        return (0, self.max_length)
-
-
-# Search view
-# simple person search view, for use with ajax calls
-def searchperson(request):
-    query = request.GET.get("q")
-    highlight = CustomHighlighter(query)
-
-    if not query or len(query) < 3:
-        return HttpResponse("")
-
-    objects = (
-        SearchQuerySet()
-        .models(models.Person)
-        .autocomplete(name_auto=query)
-        .filter_or(personnumber=query)
-    )
-
-    ret = _("No results")
-
-    if objects:
-        ret = hg.UL(
-            hg.LI(_("%s items found") % len(objects), style="margin-bottom: 20px"),
-            hg.Iterator(
-                objects,
-                "object",
-                hg.If(
-                    hg.C("object.object"),
-                    hg.LI(
-                        hg.F(
-                            lambda c, e: hg.SPAN(
-                                mark_safe(
-                                    highlight.highlight(c["object"].object.personnumber)
-                                ),
-                                style="width: 48px; display: inline-block",
-                            )
-                        ),
-                        " ",
-                        hg.F(
-                            lambda c, e: mark_safe(
-                                highlight.highlight(
-                                    c["object"].object.search_index_snippet()
-                                )
-                            )
-                        ),
-                        style="cursor: pointer; padding: 8px 0;",
-                        onclick=hg.BaseElement(
-                            "document.location = '",
-                            hg.F(
-                                lambda c, e: reverse_model(
-                                    c["object"].object,
-                                    "edit",
-                                    kwargs={"pk": c["object"].object.pk},
-                                )
-                            ),
-                            "'",
-                        ),
-                        onmouseenter="this.style.backgroundColor = 'lightgray'",
-                        onmouseleave="this.style.backgroundColor = 'initial'",
-                    ),
-                ),
-            ),
-        )
-    return HttpResponse(
-        hg.DIV(
-            ret,
-            _class="raised",
-            style="margin-bottom: 1rem; padding: 16px 0 48px 48px; background-color: #fff",
-        ).render({})
-    )
